requirements.txt
openai
genanki
azure.cognitiveservices.speech